[Refactor] #64 out 하위 디렉토리 수정
