+++ user-service/src/main/java/com/ticketrush/boundedcontext/user/app/usecase/SignupUseCase.java
@@ -5,9 +5,9 @@ import com.ticketrush.boundedcontext.user.app.dto.response.SignupResponse;
 import com.ticketrush.boundedcontext.user.domain.entity.User;
 import com.ticketrush.boundedcontext.user.domain.entity.UserAccount;
 import com.ticketrush.boundedcontext.user.domain.types.UserRole;
-import com.ticketrush.boundedcontext.user.out.EmailVerificationClient;
-import com.ticketrush.boundedcontext.user.out.UserAccountRepository;
-import com.ticketrush.boundedcontext.user.out.UserRepository;
+import com.ticketrush.boundedcontext.user.out.apiclient.EmailVerificationClient;
+import com.ticketrush.boundedcontext.user.out.repository.UserAccountRepository;
+import com.ticketrush.boundedcontext.user.out.repository.UserRepository;
 import com.ticketrush.global.exception.BusinessException;
 import com.ticketrush.global.status.ErrorStatus;
 import java.util.regex.Pattern;
