+++ user-service/src/main/java/com/ticketrush/boundedcontext/user/app/usecase/SocialLoginUseCase.java
@@ -6,8 +6,8 @@ import com.ticketrush.boundedcontext.user.app.mapper.SocialMapper;
 import com.ticketrush.boundedcontext.user.domain.entity.SocialAccount;
 import com.ticketrush.boundedcontext.user.domain.entity.User;
 import com.ticketrush.boundedcontext.user.domain.types.SocialProvider;
-import com.ticketrush.boundedcontext.user.out.SocialAccountRepository;
-import com.ticketrush.boundedcontext.user.out.UserRepository;
+import com.ticketrush.boundedcontext.user.out.repository.SocialAccountRepository;
+import com.ticketrush.boundedcontext.user.out.repository.UserRepository;
 import com.ticketrush.global.exception.BusinessException;
 import com.ticketrush.global.status.ErrorStatus;
 import java.util.Optional;
