+++ user-service/src/test/java/com/ticketrush/user/app/usecase/SignupUseCaseTest.java
@@ -12,9 +12,9 @@ import com.ticketrush.boundedcontext.user.app.dto.response.SignupResponse;
 import com.ticketrush.boundedcontext.user.app.usecase.SignupUseCase;
 import com.ticketrush.boundedcontext.user.domain.entity.User;
 import com.ticketrush.boundedcontext.user.domain.entity.UserAccount;
-import com.ticketrush.boundedcontext.user.out.EmailVerificationClient;
-import com.ticketrush.boundedcontext.user.out.UserAccountRepository;
-import com.ticketrush.boundedcontext.user.out.UserRepository;
+import com.ticketrush.boundedcontext.user.out.apiclient.EmailVerificationClient;
+import com.ticketrush.boundedcontext.user.out.repository.UserAccountRepository;
+import com.ticketrush.boundedcontext.user.out.repository.UserRepository;
 import com.ticketrush.global.exception.BusinessException;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
